requirements.txt
ultralytics==8.3.122
roboflow==1.1.63
pillow==11.2.1
matplotlib==3.10.1